refactoring item component and excluding unnecessary components
