+++ sizebayfront/src/pages/Pending/index.js
@@ -5,7 +5,7 @@ import React, { useContext, useState, useEffect } from 'react';
 import { itemContext } from '../../contexts/ItemContext';
 
 import Item from '../../components/Item';
-import { ItemsArea } from '../../components/ItemBox/styles';
+import { ItemsArea } from '../../components/Item/styles';
 
 export default function Pending() {
   const { itemsCollection, isSearch, searchText } = useContext(itemContext);
